Add function to store cookies

+++ shop/models.py
@@ -13,7 +13,7 @@ class Buyer(models.Model):
 
 class Chocolate(models.Model):
     name = models.CharField(max_length=255)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=6, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
